feat: better rectangle rendering, transparency
- Introduced transparency by adding a new TransparencyStack to the
renderer. The UI draw function can use this stack to push transparency,
which will be used by the PushRect function. The UI draw function must
also Pop this transparency from the stack once it finished with the
element.
- Use a rectangle SDF to create rounded borders (with color) and
smoothstep for antialiasing. Also setup the groudwork for supporting
gradients, left a comment with some notes on how we should probably
implement this.